+++ src/types.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.8.0;
 
 struct IndexAsset {
     address asset;
-    uint112 weightPercentage; // With 4 decimals, e.g. 500000 = 50%
+    uint128 weightPercentage; // With 4 decimals, e.g. 500000 = 50%
     address priceFeed;
 }
 
@@ -11,8 +11,8 @@ struct InitStateCache {
     uint256 priceAsset0;
     uint256 priceAsset1;
     uint256 priceUsdc;
-    uint112 initialAsset0Reserve;
-    uint112 initialAsset1Reserve;
+    uint128 initialAsset0Reserve;
+    uint128 initialAsset1Reserve;
     uint256 asset0UsdValue;
     uint256 asset1UsdValue;
     uint256 totalAssetUsdValue;
